+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Changes to Calva.
   - Adds API function `repl.listSessions()`
   - Updates default connect sequences for *Joyride*, *Babashka*, *nbb*, and *Basilisp*, to use custom `replSessionNames`, and configured default routings with `replSessionFilePatterns`.
   - Removed deprecated `calva.useLegacyReplWindowPath` setting. The REPL window is now always located at `.calva/repl.calva-repl`.
+- Adds a new connect sequence: **Clojure projectless**, for jack-in and connect without a `deps.edn`, etcetera needed.
 
 ## [2.0.540] - 2025-11-06